LPS-135720 Move util/dom to dynamic-data-mapping-form-web
- https://issues.liferay.com/browse/LPS-135720
