+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/components/FormInfo.es.js
@@ -15,14 +15,11 @@
 import ClayForm from '@clayui/form';
 import ClayLayout from '@clayui/layout';
 import {useConfig, useForm, useFormState} from 'data-engine-js-components-web';
-import {
-	isKeyInSet,
-	isModifyingKey,
-} from 'dynamic-data-mapping-form-builder/js/util/dom.es';
 import {autoSize} from 'frontend-js-web';
 import React, {useEffect, useRef} from 'react';
 
 import {EVENT_TYPES} from '../eventTypes.es';
+import {isKeyInSet, isModifyingKey} from '../util/dom';
 
 const isForbiddenKey = (event, limit) => {
 	const charCode = event.which ? event.which : event.keyCode;
