+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/components/share-form/Link.es.js
@@ -17,8 +17,9 @@ import ClayForm, {ClayInput} from '@clayui/form';
 import ClayIcon from '@clayui/icon';
 import classNames from 'classnames';
 import ClipboardJS from 'clipboard';
-import {selectText} from 'dynamic-data-mapping-form-builder/js/util/dom.es';
 import React, {useCallback, useEffect, useRef, useState} from 'react';
+
+import {selectText} from '../../util/dom';
 
 const useClipboardJS = (onSuccess) => {
 	useEffect(() => {
